acne_ai_project/requirements.txt
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.9.0
numpy==1.24.0
pandas==2.1.0
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.0
Pillow==10.1.0
albumentations==1.3.1
ultralytics==8.1.0
flask==3.0.0
flask-cors==4.0.0
tensorflowjs==4.15.0
tf2onnx==1.16.0
mediapipe==0.10.0
grad-cam==1.4.8
tqdm==4.66.0
pyyaml==6.0.1
requests==2.31.0
